refact: styled container -> for settings/messages
