+++ lib/ui/styled_components/styled_components.dart
@@ -3,3 +3,4 @@ export 'styled_dialog.dart';
 export 'styled_icon_button.dart';
 export 'styled_protocols_year.dart';
 export 'styled_ribbon_stack.dart';
+export 'styled_selectable_container.dart';
