+++ lib/ui/views/nav_dialogs/dialogs.dart
@@ -2,4 +2,3 @@ export 'about_dialog.dart';
 export 'messages_dialog.dart';
 export 'settings_dialog/settings_dialog.dart';
 export 'share_dialog.dart';
-export 'version_dialog.dart';
